+++ bld/f77/wfc/c/csutls.c
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2020 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -68,7 +68,7 @@ static  const STMT  CSWords[] = {
 };
 
 
-csnode  *NewCSNode( uint label_len )
+csnode  *NewCSNode( size_t label_len )
 {
 // Allocate a new "csnode".
 
@@ -123,7 +123,7 @@ void    AddCSNode( byte typ )
     csnode      *new_cs_node;
     itnode      *label;
     char        *label_ptr;
-    uint        label_len;
+    size_t      label_len;
 
     if( typ == CS_REMOTEBLOCK ) {
         label = CITNode;
@@ -138,7 +138,7 @@ void    AddCSNode( byte typ )
     CSHead->block = ++BlockNum;
     label_ptr = &CSHead->label;
     memcpy( label_ptr, label->opnd, label_len );
-    label_ptr[ label_len ] = NULLCHAR;
+    label_ptr[label_len] = NULLCHAR;
 }
 
 void DelCSNode(void)
