+++ bld/f77/wfc/c/declare.c
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2017 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -160,7 +160,7 @@ sym_id  FieldDecl( void ) {
 }
 
 
-TYPE    MapTypes( TYPE typ, uint size ) {
+TYPE    MapTypes( TYPE typ, size_t size ) {
 //======================================
 
 // Given a type and size, return an equivalent type.
