+++ bld/f77/wfc/c/fcdata.c
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2020 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -1404,7 +1404,7 @@ void    DtFieldSubstring( void ) {
 }
 
 
-static  void    InitArr( act_dim_list *dim, TYPE typ, uint size ) {
+static  void    InitArr( act_dim_list *dim, TYPE typ, size_t size ) {
 //=================================================================
 
 // Data initialize an array.
