+++ bld/f77/wfc/c/fcrtns.c
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2020 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -80,7 +80,7 @@ call_handle     InitCall( RTCODE rtn_id ) {
     sym_id      sym;
     rt_rtn      *rt_entry;
     byte        typ;
-    uint        name_len;
+    size_t      name_len;
     char        *ptr;
 
     rt_entry = &RtnTab[ rtn_id ];
