+++ bld/f77/wfc/c/fmtinit.c
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2020 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -52,8 +52,8 @@
 #include "cfmttab.h"
 
 
-static  void    FInit( uint fmt_length, char *fmt_string )
-//========================================================
+static  void    FInit( size_t fmt_length, char *fmt_string )
+//==========================================================
 {
     Fmt_start = fmt_string;
     Fmt_end = fmt_string + fmt_length;
@@ -82,8 +82,8 @@ static  void    FFinish( void )
     }
 }
 
-void    FScan( uint fmt_length, char *fmt_string, cs_label fmt_label )
-//====================================================================
+void    FScan( size_t fmt_length, char *fmt_string, cs_label fmt_label )
+//======================================================================
 // FORMAT statement parsing (only compile-time).
 {
     FInit( fmt_length, fmt_string );
