+++ bld/f77/wfc/c/rstnlist.c
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2019 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -39,8 +39,8 @@
 #include "cgmagic.h"
 
 
-static  sym_id  AddNameList( const char *name, uint length )
-//==========================================================
+static  sym_id  AddNameList( const char *name, size_t length )
+//============================================================
 // Add a symbol table entry to the symbol table. Return a pointer to the
 // new symbol table entry.
 {
@@ -55,8 +55,8 @@ static  sym_id  AddNameList( const char *name, uint length )
 }
 
 
-sym_id  FindNameList( const char *name, uint len )
-//================================================
+sym_id  FindNameList( const char *name, size_t len )
+//==================================================
 // Search symbol table for given name.
 {
     sym_id      head;
