+++ bld/f77/wfc/h/auxlook.h
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2020 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -30,6 +30,6 @@
 ****************************************************************************/
 
 
-extern aux_info     *AuxLookupName( const char *name, uint name_len );
-extern aux_info     *AuxLookupAdd( const char *name, uint name_len );
+extern aux_info     *AuxLookupName( const char *name, size_t name_len );
+extern aux_info     *AuxLookupAdd( const char *name, size_t name_len );
 extern aux_info     *AuxLookup( sym_id sym );
