+++ bld/f77/wfc/h/fhash.h
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2020 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -37,7 +37,7 @@ typedef struct hash_entry {
     sym_id      h_tail;
 } hash_entry;
 
-extern unsigned     CalcHash( const char *, unsigned );
+extern unsigned     CalcHash( const char *, size_t );
 extern void         HashInsert( hash_entry *, unsigned, sym_id *, sym_id );
 
 extern unsigned     HashValue;
