+++ bld/f77/wfc/h/impltab.h
@@ -2,6 +2,7 @@
 *
 *                            Open Watcom Project
 *
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -31,5 +32,5 @@
 
 extern TYPE    ImplType( char chr );
 extern uint    ImplSize( char chr );
-extern bool    SetImplType( char chr1, char chr2, TYPE typ, uint size );
+extern bool    SetImplType( char chr1, char chr2, TYPE typ, size_t size );
 extern void    InitImplTab( void );
