+++ bld/f77/wfc/h/rstutils.h
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2020 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -32,8 +32,8 @@
 
 extern char    *STExtractName( sym_id sym, char *buff );
 extern char    *STGetName( sym_id sym, char *buff );
-extern uint    AllocName( uint length );
-extern sym_id  STAdd( const char *name, int length );
+extern size_t  AllocName( size_t length );
+extern sym_id  STAdd( const char *name, size_t length );
 extern sym_id  STFree( sym_id sym );
 extern void    FreeSFHeader( sym_id sym );
 extern void    CkSymDeclared( sym_id sym );
