+++ bld/f77/wfc/h/types.h
@@ -2,6 +2,7 @@
 *
 *                            Open Watcom Project
 *
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -38,4 +39,4 @@ extern uint         TypeSize( TYPE typ );
 extern uint         StorageSize( TYPE typ );
 extern intstar4     ITIntValue( itnode *it );
 extern char         *TypeKW( TYPE typ );
-extern PTYPE        ParmType( TYPE typ, uint size );
+extern PTYPE        ParmType( TYPE typ, size_t size );
